sdk/python/kubeflow/storage_init_container/requirements.txt
torch==2.1.1
torchvision==0.16.1
torchaudio==2.1.1
einops==0.7.0
transformers_stream_generator==0.0.4
